# docs/club-management-openapi.yaml
openapi: 3.0.3
info:
  title: Club Management System API
  version: 1.0.0
  description: |
    社团管理系统 OpenAPI 文档，覆盖认证、用户、社团、部门、岗位、招新、申请、审核、面试、成员、退社、RBAC、通知与日志模块。
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /clubs/{clubId}/departments:
    get:
      tags: [Departments]
      summary: 查询社团部门列表
      parameters:
        - $ref: '#/components/parameters/ClubId'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Department'
    post:
      tags: [Departments]
      summary: 创建部门
      parameters:
        - $ref: '#/components/parameters/ClubId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /departments/{departmentId}:
    get:
      tags: [Departments]
      summary: 获取部门详情
      parameters:
        - $ref: '#/components/parameters/DepartmentId'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Department'
    patch:
      tags: [Departments]
      summary: 更新部门
      parameters:
        - $ref: '#/components/parameters/DepartmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDepartmentRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags: [Departments]
      summary: 删除部门
      parameters:
        - $ref: '#/components/parameters/DepartmentId'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /notifications:
    get:
      tags: [Notifications]
      summary: 查询当前用户通知列表
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
    post:
      tags: [Notifications]
      summary: 发送通知
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /notifications/{notificationId}/read:
    post:
      tags: [Notifications]
      summary: 标记通知已读
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ClubId:
      name: clubId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    DepartmentId:
      name: departmentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    NotificationId:
      name: notificationId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
        traceId:
          type: string
...
    Department:
      type: object
      properties:
        id:
          type: integer
          format: int64
        clubId:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        managerUserId:
          type: integer
          format: int64
    CreateDepartmentRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        managerUserId:
          type: integer
          format: int64
    UpdateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        managerUserId:
          type: integer
          format: int64
...
    Notification:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        type:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateNotificationRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
        receiverUserIds:
          type: array
          items:
            type: integer
            format: int64
...